Restructure codebase even further
